[MODULES] update modules
