+++ modules/nf-core/cooler/cload/meta.yml
@@ -1,0 +1,53 @@
+name: cooler_cload
+description: Create a cooler from genomic pairs and bins
+keywords:
+  - cool
+tools:
+  - cooler:
+      description: Sparse binary format for genomic interaction matrices
+      homepage: https://open2c.github.io/cooler/
+      documentation: https://cooler.readthedocs.io/en/latest/index.html
+      tool_dev_url: https://github.com/open2c/cooler
+      doi: "10.1093/bioinformatics/btz540"
+      licence: ["BSD-3-clause"]
+
+input:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - pairs:
+      type: file
+      description: Path to contacts (i.e. read pairs) file.
+  - index:
+      type: file
+      description: Path to index file of the contacts.
+  - cool_bin:
+      type: value
+      description: Bins size in bp
+  - chromsizes:
+      type: file
+      description: Path to a chromsizes file.
+
+output:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test', single_end:false ]
+  - version:
+      type: file
+      description: File containing software version
+      pattern: "versions.yml"
+  - cool:
+      type: file
+      description: Output COOL file path
+      pattern: "*.cool"
+  - cool_bin:
+      type: value
+      description: Bins size in bp
+
+authors:
+  - "@jianhong"
+  - "@muffato"
