+++ modules/nf-core/cooler/makebins/main.nf
@@ -1,31 +1,30 @@
-process COOLER_ZOOMIFY {
-    tag "$meta.id"
-    label 'process_high'
+process COOLER_MAKEBINS {
+    tag "${meta.id}}"
+    label 'process_low'
 
-    conda (params.enable_conda ? "bioconda::cooler=0.8.11" : null)
+    conda "bioconda::cooler=0.8.11"
     container "${ workflow.containerEngine == 'singularity' && !task.ext.singularity_pull_docker_container ?
-        'https://depot.galaxyproject.org/singularity/cooler:0.8.11--pyh3252c3a_0' :
+        'https://depot.galaxyproject.org/singularity/cooler:0.8.11--pyh3252c3a_0':
         'quay.io/biocontainers/cooler:0.8.11--pyh3252c3a_0' }"
 
     input:
-    tuple val(meta), path(cool)
+    tuple val(meta), path(chromsizes), val(cool_bin)
 
     output:
-    tuple val(meta), path("*.mcool"), emit: mcool
-    path "versions.yml"             , emit: versions
+    tuple val(meta), path("*.bed"), emit: bed
+    path "versions.yml"           , emit: versions
 
     when:
     task.ext.when == null || task.ext.when
 
     script:
-    def args = task.ext.args ?: ''
+    def args   = task.ext.args   ?: ''
     def prefix = task.ext.prefix ?: "${meta.id}"
     """
-    cooler zoomify \\
+    cooler makebins \\
         $args \\
-        -n $task.cpus \\
-        -o ${prefix}.mcool \\
-        $cool
+        ${chromsizes} \\
+        ${cool_bin} > ${prefix}.bed
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
