+++ modules/nf-core/cooler/zoomify/meta.yml
@@ -5,7 +5,7 @@ keywords:
 tools:
   - cooler:
       description: Sparse binary format for genomic interaction matrices
-      homepage: https://cooler.readthedocs.io/en/latest/index.html
+      homepage: https://open2c.github.io/cooler/
       documentation: https://cooler.readthedocs.io/en/latest/index.html
       tool_dev_url: https://github.com/open2c/cooler
       doi: "10.1093/bioinformatics/btz540"
